requirements.txt
requests==2.28.1 
gradio
openai==0.28 
python-dotenv==1.0.0 
httpx==0.21.1 
httpcore==0.14.7